feat: Add cluster-secret application deployment
This commit adds the YAML file for deploying the cluster-secret application as a Kubernetes deployment. The deployment is configured to use the latest version of the cluster-secret chart from the repository https://charts.clustersecret.io. The application is deployed in the "clustersecret" namespace and is part of the "default" project. The sync policy is set to automatically prune and self-heal the resources, with various sync options configured. The retry limit is set to 5 with a backoff duration of 5 seconds. The revision history limit is set to 3.